pkg/registry: move `stack` to its own file
